--- xray/commands/audit/jas/testutils.go
@@ -1,41 +1,0 @@
-package jas
-
-import (
-	rtutils "github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
-	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
-	"github.com/jfrog/jfrog-client-go/xray/services"
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-var FakeServerDetails = config.ServerDetails{
-	Url:      "platformUrl",
-	Password: "password",
-	User:     "user",
-}
-
-var FakeBasicXrayResults = []services.ScanResponse{
-	{
-		ScanId: "scanId_1",
-		Vulnerabilities: []services.Vulnerability{
-			{IssueId: "issueId_1", Technology: coreutils.Pipenv.ToString(),
-				Cves:       []services.Cve{{Id: "testCve1"}, {Id: "testCve2"}, {Id: "testCve3"}},
-				Components: map[string]services.Component{"issueId_1_direct_dependency": {}, "issueId_3_direct_dependency": {}}},
-		},
-		Violations: []services.Violation{
-			{IssueId: "issueId_2", Technology: coreutils.Pipenv.ToString(),
-				Cves:       []services.Cve{{Id: "testCve4"}, {Id: "testCve5"}},
-				Components: map[string]services.Component{"issueId_2_direct_dependency": {}, "issueId_4_direct_dependency": {}}},
-		},
-	},
-}
-
-func InitJasTest(t *testing.T, workingDirs ...string) (*JasScanner, func()) {
-	assert.NoError(t, rtutils.DownloadAnalyzerManagerIfNeeded())
-	scanner, err := NewJasScanner(workingDirs, &FakeServerDetails)
-	assert.NoError(t, err)
-	return scanner, func() {
-		assert.NoError(t, scanner.ScannerDirCleanupFunc())
-	}
-}
